Decompose monolithic jaconv.py into domain-specific modules
Refactor the 959-line jaconv.py into 6 focused modules:

- helpers.py: Shared internal helpers (_exclude_ignorechar, _convert, _translate)
- kana_convert.py: Hiragana/Katakana conversion (hira2kata, hira2hkata, kata2hira, enlarge_smallkana)
- width_convert.py: Half-width/Full-width conversion (h2z, z2h + aliases)
- normalize_convert.py: Unicode normalization (normalize)
- romaji_convert.py: Kana↔Alphabet conversion (kana2alphabet, alphabet2kana, kata2alphabet, alphabet2kata)
- julius_convert.py: Julius phoneme format conversion (hiragana2julius)

jaconv.py becomes a thin facade that re-exports all public functions.
Public API is unchanged — all 17 functions work identically.

Verified with:
- All 10 existing unit tests pass
- Regression testing via Regrets fingerprint tool (14 clusters, all GREEN)
- Raw output comparison: 45 input/output pairs identical before and after refactor
